rewrite tokenizer
