--- internal/experimental_tokenizer/testfiles/nested/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-export { default as func } from "./test";
-export * from "./test2";
-
-export const IndexComponent = () => {};
