+++ internal/tokenizer/testfiles/mdn-export-examples.js
