+++ internal/tokenizer/testfiles/mdn-import-examples.js
