+++ internal/tokenizer/tokenizer.go
@@ -1,8 +1,3 @@
-/*
-Regex seems to mostly work for this but is a little challenging when it comes to handling comments
-especially since Go does not support lookbehind regex (for performance reasons, apparently).
-I didn't think it would be crazy difficult to write an import lexer so that's what this is.
-*/
 package tokenizer
 
 import (
@@ -10,25 +5,22 @@ import (
 	"os"
 	"path/filepath"
 	"slices"
-	"strings"
 	"unicode"
 )
 
-const (
-	REQUIRE_LENGTH = 7
-	IMPORT_LENGTH  = 6
-	EXPORT_LENGTH  = 6
-)
-
 type Tokenizer struct {
+	// points to index of current `char`
 	currentIndex int
-	fileRunes    []rune
-	imports      map[string][]string
-	reExports    []string
-	reExportMap  map[string]string
-	exports      []string
-	callDir      string
-	initPath     string
+	// points to index of next character to read
+	readIndex   int
+	char        rune
+	fileRunes   []rune
+	imports     map[string][]string
+	reExports   []string
+	reExportMap map[string]string
+	exports     []string
+	callDir     string
+	initPath    string
 }
 
 func NewTokenizerFromFile(initPath string) (*Tokenizer, error) {
@@ -42,7 +34,8 @@ func NewTokenizerFromFile(initPath string) (*Tokenizer, error) {
 
 func New(fileString, initPath string) *Tokenizer {
 	t := Tokenizer{
-		currentIndex: 0,
+		currentIndex: -1,
+		readIndex:    0,
 		fileRunes:    []rune(fileString),
 		imports:      make(map[string][]string, 0),
 		reExports:    []string{},
@@ -51,185 +44,142 @@ func New(fileString, initPath string) *Tokenizer {
 		callDir:      filepath.Dir(initPath),
 		initPath:     initPath,
 	}
+	t.readChar()
 	return &t
 }
 
-// Something to keep in mind: this will at least miss *some* edge cases.
-// e.g. nested requires `require(require('./pathToRealImport'))`
-// Supporting this case seems challenging and I don't think it's currently worth my effort
-// Reads all import paths from a file in one pass and returns them in an array
 func (t *Tokenizer) Tokenize() FileToken {
 	if len(t.fileRunes) < 1 {
 		return FileToken{}
 	}
 
-	for t.currentIndex < t.end() {
-		char := t.current()
-		switch char {
-		case 'i':
-			t.tokenizeImport()
-		case 'r':
-			t.tokenizeRequire()
-		case 'e':
-			t.tokenizeExport()
-		case '/':
-			t.skipComment(t.peek())
+	for t.char != 0 {
+		if t.char == 'i' || t.char == 'r' || t.char == 'e' {
+			switch token := t.readIdentifier(); token {
+			case "import":
+				t.readImport()
+			case "require":
+				t.readRequire()
+			case "export":
+				t.readExport()
+			}
+		} else if t.char == '/' {
+			// comments could contain keywords in them but should not be parsed as imports / exports
+			t.skipComment(true)
+		} else {
+			// avoids reading char if a previous read got us to EOF
+			t.readChar()
 		}
-		t.advanceChars()
 	}
+
 	return FileToken{
 		FilePath:    t.initPath,
 		Imports:     t.imports,
 		ReExports:   t.reExports,
-		ReExportMap: t.reExportMap,
 		Exports:     t.exports,
+		ReExportMap: t.reExportMap,
 	}
 }
 
-func (t *Tokenizer) skipComment(peekChar rune) {
-	switch peekChar {
-	case '/':
-		t.skipSingleLineComment()
-	case '*':
-		t.skipMultiLineComment()
-	default:
-		return
-	}
-}
+// Export cases: https://developer.mozilla.org/en-US/docs/web/javascript/reference/statements/export
+func (t *Tokenizer) readExport() {
+	var identifiers []string
+	isReExport := false
+	haveSeenLeftBrace := false
+	// '}' is also sort of an an endChar. e.g. export { foo, bar } w/o semi-colon
+	// but it's ambiguous because for re-exports we'd need to continue. I'm not sure
+	// there's any way to handle this case besides looking ahead afterwards
+	endedOnBrace := false
+	// overwrite exported identifiers with their aliases so that they are correctly mapped to importing files
+	overwriteLastIdentifier := false
+	endChars := []rune{';', '(', '='}
 
-// Finds the first quoted string after an import keyword and adds it to imports
-// TODO: do I also need to handle identifiers inside of parens?
-func (t *Tokenizer) tokenizeImport() {
-	if t.currentIndex+IMPORT_LENGTH > t.end() {
-		return
-	}
-
-	// the word `import` could be used in a variable name so this protects agains that
-	if t.currentIndex != 0 && t.prev() != ';' && !unicode.IsSpace(t.prev()) {
-		return
-	}
-
-	if string(t.fileRunes[t.currentIndex:t.currentIndex+IMPORT_LENGTH]) != "import" {
-		return
-	}
-
-	t.advanceChars(IMPORT_LENGTH)
-	// in this case import is probably part of a variable name
-	if t.current() != '{' && t.current() != '(' && !unicode.IsSpace(t.current()) {
-		return
-	}
-
-	identifiers := t.tokenizeImportIdentifiers()
-	// This will allow some incorrect syntax to be treated as an import, e.g. import './not/a/real/import'
-	// But fixing that is low priority for now
-	for t.currentIndex < t.end() && !isQuote(t.current()) {
-		if t.current() == ';' || t.current() == ')' {
-			return
+Loop:
+	for t.char != 0 {
+		switch {
+		// this needs to come before isIdentifierEnd check because some of these chars are shared
+		case slices.Contains(endChars, t.char):
+			break Loop
+		case t.char == '{':
+			haveSeenLeftBrace = true
+			t.readChar()
+		case t.char == ':':
+			overwriteLastIdentifier = true
+			t.readChar()
+		case t.char == '/':
+			t.skipComment(false)
+		case t.char == '}':
+			endedOnBrace = true
+			break Loop
+		case isIdentifierEnd(t.char):
+			t.readChar()
+		case isQuote(t.char):
+			panic(fmt.Sprintf("Encountered a quote in an export statement that was not preceded by the `from` keyword in %q. This may be a syntax error or it could be a quoted export alias. Dependor does not currently support quoted import aliases.", t.initPath))
+		default:
+			ident := t.readIdentifier()
+			// I don't think this case can happen, but if it does, this will avoid an infinite loop
+			if len(ident) == 0 && t.char != 0 {
+				panic(fmt.Sprintf("Unexpected length 0 identifier not located at the end of the file in %q. This situation was expected to be impossible.", t.initPath))
+			}
+			switch ident {
+			case "as":
+				overwriteLastIdentifier = true
+			case "from":
+				isReExport = true
+				break Loop
+			case "interface":
+				endChars = append(endChars, '{')
+			case "const", "let", "var", "function", "function*", "class", "type":
+				continue
+			default:
+				if ident == "default" && !haveSeenLeftBrace {
+					identifiers = append(identifiers, ident)
+					break Loop
+				}
+				if overwriteLastIdentifier {
+					identifiers[len(identifiers)-1] = ident
+					overwriteLastIdentifier = false
+					continue
+				}
+				identifiers = append(identifiers, ident)
+			}
 		}
-		if t.current() == '/' {
-			// Avoids interpreting quotes inside of comments as strings
-			t.skipImportComments()
+	}
+	if endedOnBrace {
+		t.readChar()
+		t.skipAllFiller()
+		if t.char == 'f' {
+			ident := t.readIdentifier()
+			if ident == "from" {
+				isReExport = true
+			}
 		}
-		t.advanceChars()
 	}
 
-	importPath, ok := t.readImportString()
-	if !ok {
-		return
-	}
-
-	if _, exists := t.imports[importPath]; exists {
-		t.imports[importPath] = append(t.imports[importPath], identifiers...)
-		return
-	}
-
-	t.imports[importPath] = identifiers
-}
-
-func (t *Tokenizer) tokenizeRequire() {
-	if t.currentIndex+REQUIRE_LENGTH > t.end() {
-		return
-	}
-
-	if t.currentIndex != 0 && t.prev() != ';' && !unicode.IsSpace(t.prev()) {
-		return
-	}
-	// This somewhat breaks doing things in one pass but I suspect the performance drawbacks
-	// here aren't significant enough to optimize right away
-	if string(t.fileRunes[t.currentIndex:t.currentIndex+REQUIRE_LENGTH]) != "require" {
-		return
-	}
-
-	t.advanceImportChars(REQUIRE_LENGTH)
-	if t.current() != '(' {
-		return
-	}
-
-	t.advanceImportChars()
-	if !isQuote(t.current()) {
-		return
-	}
-	path, ok := t.readImportString()
-	if ok {
-		t.imports[path] = []string{}
-	}
-}
-
-func (t *Tokenizer) tokenizeExport() {
-	if t.currentIndex+EXPORT_LENGTH > t.end() {
-		return
-	}
-
-	// the word `export` could be used in a variable name so this protects against that
-	if t.currentIndex != 0 && t.prev() != ';' && !unicode.IsSpace(t.prev()) {
-		return
-	}
-
-	if string(t.fileRunes[t.currentIndex:t.currentIndex+EXPORT_LENGTH]) != "export" {
-		return
-	}
-
-	t.advanceChars(EXPORT_LENGTH)
-	// in this case export is probably part of a variable name
-	if t.current() != '{' && !unicode.IsSpace(t.current()) {
-		return
-	}
-
-	identifiers, isRegularExport := t.tokenizeExportIdentifiers()
-
-	if isRegularExport {
+	if !isReExport {
 		t.exports = append(t.exports, identifiers...)
-		return
-	}
-	// TODO: This is used multiple times and could probably be abstracted to a `findQuote` function
-	// in this case we are dealing with a re-export so we need to find the next string literal
-	for current := t.current(); t.currentIndex < t.end() && !isQuote(current); current = t.current() {
-		if current == ';' || current == ')' {
-			return
-		}
-		if current == '/' {
-			// Avoids interpreting quotes inside of comments as strings
-			t.skipImportComments()
-		}
-		t.advanceChars()
-	}
-
-	reExportPath, ok := t.readImportString()
-	if !ok {
-		return
-	}
-
-	t.reExports = append(t.reExports, reExportPath)
-	if len(identifiers) == 0 {
-		fmt.Printf("WARN: file %q has no reExported identifiers but contained a reExport. This is either a syntax error, or a bug in the tokenizer.\n", t.initPath)
 		return
 	}
 
 	if t.reExportMap == nil {
-		t.reExportMap = make(map[string]string, len(identifiers))
+		t.reExportMap = make(map[string]string, 0)
 	}
+
+	t.skipAllFiller()
+	if !isQuote(t.char) {
+		panic(fmt.Sprintf("Unexpected non-string token following the keyword `from` in %q. This is likely due to a syntax error.", t.initPath))
+	}
+	reExportPath := t.readPathString()
+
+	t.reExports = append(t.reExports, reExportPath)
+	if len(identifiers) == 0 {
+		panic(fmt.Sprintf("Unexpected re-export with zero identifiers in %q. This is likely a syntax error.", t.initPath))
+	}
+
 	// populate reExportMap with idents. If an ident is "*"
-	// save reExport path in map so that it can be populated in the parser
+	// save reExport path in map so that it can be populated
+	// later inside of the parser. For aliased *'s (e.g. export * as namespace from './file')
+	// then mapping the alias to the file path should be sufficient
 	for _, ident := range identifiers {
 		if ident == "*" {
 			t.reExportMap[reExportPath] = "*"
@@ -239,233 +189,190 @@ func (t *Tokenizer) tokenizeExport() {
 	}
 }
 
-func (t *Tokenizer) readImportString() (string, bool) {
-	var b strings.Builder
-	t.advanceChars()
-	for t.currentIndex < t.end() && !isQuote(t.current()) {
-		b.WriteRune(t.current())
-		t.advanceChars()
-	}
-
-	if t.currentIndex >= t.end() {
-		return "", false
-	}
-
-	path := b.String()
-	if isRelativePath(path) {
-		path = filepath.Join(t.callDir, path)
-	}
-	return path, true
-}
-
-func (t *Tokenizer) skipSingleLineComment() {
-	for t.currentIndex < t.end() && t.current() != '\n' {
-		t.advanceChars()
-	}
-}
-
-func (t *Tokenizer) skipMultiLineComment() {
-	for t.currentIndex+1 < t.end() && !(t.current() == '*' && t.peek() == '/') {
-		t.advanceChars()
-	}
-	t.advanceChars(2)
-}
-
-func (t *Tokenizer) skipWhitespace() {
-	for t.currentIndex < t.end() && unicode.IsSpace(t.current()) {
-		t.advanceChars()
-	}
-}
-
-func (t *Tokenizer) skipImportComments() {
-	for t.currentIndex < t.end() && t.current() == '/' {
-		t.skipComment(t.peek())
-		t.skipWhitespace()
-	}
-}
-
-func (t *Tokenizer) tokenizeImportIdentifiers() []string {
-	endChars := []rune{';', '(', '"', '\'', '`'}
-	stopChars := []rune{',', '{', '}', '/'}
-	var currentIdentifier []rune
+// Note: this can end up tokenizing invalid javascript e.g. import +not+a+valid+var from 'path';
+// We try to throw errors on these cases when possible but checking a valid javascript variable
+// is pretty complex because the spec allows many unicode characters including emojis. Instead
+// we just assume any non-whitespace character not in identifier_ends is a valid identifier char
+func (t *Tokenizer) readImport() {
 	var identifiers []string
-	isDefault := true
-	for t.currentIndex < t.end() && !slices.Contains(endChars, t.current()) {
-		if current := t.current(); slices.Contains(stopChars, current) || unicode.IsSpace(current) {
-			ident := string(currentIdentifier)
-			switch {
-			case ident == "as":
-				t.skipNextIdentifier()
-			case ident == "from":
-				break
-			case ident == "type":
-				// ignore `type` since it can be used as annotation `import type { FooType } ...`
-			case len(ident) > 0:
-				if isDefault {
-					identifiers = append(identifiers, "default")
-				} else {
-					identifiers = append(identifiers, ident)
-				}
-			}
-
-			currentIdentifier = []rune{}
-
-			switch current {
-			case '{':
-				isDefault = false
-			case '}':
-				isDefault = true
-			case '/':
-				if t.peek() == '*' || t.peek() == '/' {
-					t.skipImportComments()
-					continue
-				}
-			}
-		} else {
-			currentIdentifier = append(currentIdentifier, current)
-		}
-
-		t.advanceChars()
-	}
-
-	return identifiers
-}
-
-// returns false when 'from' token is encountered
-func (t *Tokenizer) tokenizeExportIdentifiers() ([]string, bool) {
-	// TODO: recreating and garbage collecting this on each function call is probably bad for performance
-	keywords := []string{"const", "let", "var", "function", "interface", "type"}
-	stopChars := []rune{',', '{', '}', '/'}
-	endChars := []rune{'=', ';', '('}
-
-	var currentIdentifier []rune
-	var identifiers []string
-	// When `default`` is inside of curly braces it means we are re-exporting
-	// in this case we need to return false
+	skipNextIdentifier := false
+	// used to determine if import is a default import
 	haveSeenLeftBrace := false
-	// For mapping to work correctly we need to overwrite aliased exports with their alias
-	overwriteLastExport := false
-	for t.currentIndex < t.end() && !slices.Contains(endChars, t.current()) {
-		if current := t.current(); slices.Contains(stopChars, current) || unicode.IsSpace(current) {
-			if current == '{' {
+	for t.char != 0 {
+		switch {
+		case t.char == '/':
+			t.skipComment(false)
+		case isIdentifierEnd(t.char):
+			if t.char == '{' {
 				haveSeenLeftBrace = true
 			}
-			ident := string(currentIdentifier)
+			t.readChar()
+		case t.char == ')':
+			// import() can import using a variable rather than a string I think
+			return
+		case isQuote(t.char):
+			importPath := t.readPathString()
+			t.imports[importPath] = append(t.imports[importPath], identifiers...)
+			return
+		default:
+			ident := t.readIdentifier()
+			// I don't think this case can happen, but if it does, this will avoid an infinite loop
+			if len(ident) == 0 && t.char != 0 {
+				panic(fmt.Sprintf("Unexpected length 0 identifier not located at the end of the file in %q. This situation was expected to be impossible.", t.initPath))
+			}
+
 			switch ident {
 			case "as":
-				overwriteLastExport = true
-			case "from":
-				return identifiers, false
-			case "default":
-				if !haveSeenLeftBrace {
-					return []string{"default"}, true
-				}
-				// in this case we are in a re-export
-				identifiers = addExportIdentifier(identifiers, ident, overwriteLastExport)
-				overwriteLastExport = false
-			case "interface":
-				// interfaces are of form export interface ident { //... }
-				endChars = append(endChars, '{')
-			default:
-				if !slices.Contains(keywords, ident) && len(ident) > 0 {
-					identifiers = addExportIdentifier(identifiers, ident, overwriteLastExport)
-					overwriteLastExport = false
-				}
-			}
-
-			currentIdentifier = []rune{}
-			if current == '/' && (t.peek() == '/' || t.peek() == '*') {
-				t.skipImportComments()
+				skipNextIdentifier = true
+			// Some typescript setups annotate imports of types as `import type { ... } ...`
+			case "from", "type":
 				continue
+			default:
+				if skipNextIdentifier {
+					skipNextIdentifier = false
+					continue
+				}
+				// non-bracketed imports are either default imports or namespace imports
+				if !haveSeenLeftBrace && ident != "*" {
+					ident = "default"
+				}
+				identifiers = append(identifiers, ident)
 			}
-		} else {
-			currentIdentifier = append(currentIdentifier, current)
 		}
-
-		t.advanceChars()
 	}
-
-	return identifiers, true
+	panic(fmt.Sprintf("Encountered non-terminating import statement in %q. This is likely a syntax error.", t.initPath))
 }
 
-// Assumes correct syntax. This would fail for the case `as from "./path";`
-func (t *Tokenizer) skipNextIdentifier() {
-	endChars := []rune{';', '(', ',', '{', '}'}
-	t.skipWhitespace()
-	for current := t.current(); t.currentIndex < t.end() && !slices.Contains(endChars, current) && !unicode.IsSpace(current); current = t.current() {
-		if current == '/' && (t.peek() == '*' || t.peek() == '/') {
-			t.skipImportComments()
+// skips to first non-whitespace non-comment character
+func (t *Tokenizer) skipAllFiller() {
+Loop:
+	for t.char != 0 {
+		switch {
+		case t.char == '/':
+			t.skipComment(false)
+		case unicode.IsSpace(t.char):
+			t.skipWhitespace()
+		default:
+			break Loop
 		}
-		t.advanceChars()
 	}
 }
 
-func (t *Tokenizer) peek() rune {
-	// TODO: Refactor this to be less sloppy
-	if t.currentIndex+1 >= t.end() {
-		return 0
+func (t *Tokenizer) readRequire() {
+	for t.char != 0 {
+		switch {
+		case t.char == ')':
+			return
+		case t.char == '/':
+			t.skipComment(false)
+		case isQuote(t.char):
+			requirePath := t.readPathString()
+			t.imports[requirePath] = []string{}
+			return
+		default:
+			t.readChar()
+		}
 	}
-	return t.fileRunes[t.currentIndex+1]
+	panic(fmt.Sprintf("Encountered a non-terminating require statement in %q. This is likely a syntax error.", t.initPath))
 }
 
-func (t *Tokenizer) current() rune {
-	if t.currentIndex >= t.end() {
-		return 0
+// skips to first non-whitespace character
+func (t *Tokenizer) skipWhitespace() {
+	for unicode.IsSpace(t.char) {
+		t.readChar()
 	}
-	return t.fileRunes[t.currentIndex]
 }
 
-// This is useful for import to determine if it is part of a variable name
-// It's not strictly necessary because we could use the peek() to accomplish
-// this as well but for now it's making my life easier.
-func (t *Tokenizer) prev() rune {
-	if t.currentIndex < 1 {
-		return 0
+// we try to panic on syntax errors when it's reasonable to do so rather than give incorrect output:
+// when slashes show up in import / export statements they should be 1) part of a string or 2) part of a comment.
+// when slashes show up outside of import / export statements then they are valid JavaScript and we should not panic
+func (t *Tokenizer) skipComment(standaloneSlashIsSafe bool) {
+	t.readChar()
+	switch t.char {
+	case '/':
+		t.skipSingleLineComment()
+	case '*':
+		t.skipMultiLineComment()
+	default:
+		// Unless this function is called at the top level, standalone slashes are invalid syntax.
+		if !standaloneSlashIsSafe {
+			panic(fmt.Sprintf("Error: tokenizer came across an unexpected '/' character in an import or export statement that is not part of an import string or comment. This occurred in %q and is likely due to a syntax error.\n", t.initPath))
+		}
 	}
-	return t.fileRunes[t.currentIndex-1]
+}
+
+// skips to first character that is not part of a single line comment
+func (t *Tokenizer) skipSingleLineComment() {
+	for t.char != 0 && t.char != '\n' && t.char != '\r' {
+		t.readChar()
+	}
+	if t.char != 0 {
+		t.readChar()
+	}
+}
+
+// Skips to first character that is not part of a multilinie comment
+func (t *Tokenizer) skipMultiLineComment() {
+	t.readChar()
+	for t.char != 0 {
+		if t.char == '*' && t.peek() == '/' {
+			break
+		}
+		t.readChar()
+	}
+	t.readChar()
+	t.readChar()
+}
+
+func (t *Tokenizer) readIdentifier() string {
+	start := t.currentIndex
+	for t.char != 0 && !isIdentifierEnd(t.char) && !isQuote(t.char) {
+		t.readChar()
+	}
+	return string(t.fileRunes[start:t.currentIndex])
+}
+
+func (t *Tokenizer) readPathString() string {
+	// should be starting on a quote so we need to advance to first nonquote
+	t.readChar()
+	start := t.currentIndex
+	for t.currentIndex < t.end() {
+		if isQuote(t.char) {
+			break
+		}
+		t.readChar()
+	}
+
+	if t.currentIndex == t.end() {
+		// Panic here since syntax errors like this could silently cause the
+		// resultant import graph to be incorrect.
+		panic(fmt.Sprintf("Error: tokenizer came across a non-terminating string in %q. This is likely a syntax error.\n", t.initPath))
+	}
+
+	pathString := string(t.fileRunes[start:t.currentIndex])
+	if isRelativePath(pathString) {
+		pathString = filepath.Join(t.callDir, pathString)
+	}
+	return pathString
+}
+
+func (t *Tokenizer) readChar() {
+	if t.readIndex >= t.end() {
+		t.char = 0
+		return
+	}
+	t.char = t.fileRunes[t.readIndex]
+	t.readIndex++
+	t.currentIndex++
 }
 
 func (t *Tokenizer) end() int {
 	return len(t.fileRunes)
 }
 
-// Advances chars and skips whitespace and comments
-func (t *Tokenizer) advanceImportChars(args ...int) {
-	step := 1
-	if len(args) > 0 && args[0] > 0 {
-		step = args[0]
+func (t *Tokenizer) peek() rune {
+	if t.readIndex == len(t.fileRunes) {
+		return 0
 	}
-	t.advanceChars(step)
-	t.skipWhitespace()
-	t.skipImportComments()
-}
-
-// Supports optional int to advance by more than one. Subsequent ints are ignored.
-func (t *Tokenizer) advanceChars(args ...int) {
-	step := 1
-	if len(args) > 0 && args[0] > 0 {
-		step = args[0]
-	}
-	t.currentIndex += step
-}
-
-func isQuote(c rune) bool {
-	return c == '\'' || c == '"' || c == '`'
-}
-
-func isRelativePath(path string) bool {
-	return strings.HasPrefix(path, ".")
-}
-
-// Performance note (because I'm bad at Go): Slices are passed by value but a slice is
-// simply a header that gives information about an underlying array (e.g. pointer to array
-// plus start and stop indices). So reassigning a slice is not nearly as expensive as
-// copying and reassigning an array in languages like JavaScript
-func addExportIdentifier(identifiers []string, ident string, overwrite bool) []string {
-	if overwrite && len(identifiers) > 0 {
-		identifiers[len(identifiers)-1] = ident
-	} else {
-		identifiers = append(identifiers, ident)
-	}
-	return identifiers
+	return t.fileRunes[t.readIndex]
 }
